+++ compiler/ml/printtyp.ml
@@ -146,6 +146,10 @@ let string_of_label = function
   | Labelled s -> s
   | Optional s -> "?" ^ s
 
+let string_of_arity = function
+  | None -> ""
+  | Some arity -> string_of_int arity
+
 let visited = ref []
 let rec raw_type ppf ty =
   let ty = safe_repr [] ty in
@@ -159,9 +163,10 @@ and raw_type_list tl = raw_list raw_type tl
 
 and raw_type_desc ppf = function
   | Tvar name -> fprintf ppf "Tvar %a" print_name name
-  | Tarrow (l, t1, t2, c) ->
-    fprintf ppf "@[<hov1>Tarrow(\"%s\",@,%a,@,%a,@,%s)@]" (string_of_label l)
-      raw_type t1 raw_type t2 (safe_commu_repr [] c)
+  | Tarrow (l, t1, t2, c, a) ->
+    fprintf ppf "@[<hov1>Tarrow(\"%s\",@,%a,@,%a,@,%s,@,%s)@]"
+      (string_of_label l) raw_type t1 raw_type t2 (safe_commu_repr [] c)
+      (string_of_arity a)
   | Ttuple tl -> fprintf ppf "@[<1>Ttuple@,%a@]" raw_type_list tl
   | Tconstr (p, tl, abbrev) ->
     fprintf ppf "@[<hov1>Tconstr(@,%a,@,%a,@,%a)@]" path p raw_type_list tl
@@ -501,7 +506,7 @@ let rec mark_loops_rec visited ty =
     let visited = px :: visited in
     match ty.desc with
     | Tvar _ -> add_named_var ty
-    | Tarrow (_, ty1, ty2, _) ->
+    | Tarrow (_, ty1, ty2, _, _) ->
       mark_loops_rec visited ty1;
       mark_loops_rec visited ty2
     | Ttuple tyl -> List.iter (mark_loops_rec visited) tyl
@@ -582,7 +587,7 @@ let rec tree_of_typexp sch ty =
         let non_gen = is_non_gen sch ty in
         let name_gen = if non_gen then new_weak_name ty else new_name in
         Otyp_var (non_gen, name_of_type name_gen ty)
-      | Tarrow (l, ty1, ty2, _) ->
+      | Tarrow (l, ty1, ty2, _, _) ->
         let pr_arrow l ty1 ty2 =
           let lab = string_of_label l in
           let t1 =