+++ compiler/ml/tast_iterator.ml
@@ -295,7 +295,7 @@ let typ sub {ctyp_desc; ctyp_env; _} =
   match ctyp_desc with
   | Ttyp_any -> ()
   | Ttyp_var _ -> ()
-  | Ttyp_arrow (_, ct1, ct2) ->
+  | Ttyp_arrow (_, ct1, ct2, _) ->
     sub.typ sub ct1;
     sub.typ sub ct2
   | Ttyp_tuple list -> List.iter (sub.typ sub) list